+++ test/generators.test.inc.js
@@ -4,67 +4,29 @@
 // --------------------
 
 // modules
-var chai = require('chai'),
-	expect = chai.expect,
-	Promise = global.Promise || require('bluebird'),
-	series = require('../lib/');
+var Promise = global.Promise || require('bluebird');
 
-// init
-chai.config.includeStack = true;
+// imports
+var runTests = require('./runTests.inc');
 
-// tests
+// run tests
 
-/* jshint expr: true */
 /* jshint esnext: true */
-/* global describe, it */
 
-describe('Generator', function() {
-	it('runs serially', function() {
-		var order = [];
-		var fn = series(function*(i) {
-			order.push('start' + i);
-			yield Promise.resolve();
-			order.push('end' + i);
-			return i * 10;
-		});
+var order = [];
 
-		var res = [1, 2, 3].map(fn);
-		order.push('sync');
-
-		return res[2].then(function(res) {
-			expect(res).to.equal(30);
-			expect(order).to.deep.equal(['start1', 'sync', 'end1', 'start2', 'end2', 'start3', 'end3']);
-		});
-	});
-
-	it('delays first iteration with option {immediate: false}', function() {
-		var order = [];
-		var fn = series(function*(i) {
-			order.push('start' + i);
-			yield Promise.resolve();
-			order.push('end' + i);
-			return i * 10;
-		}, {immediate: false});
-
-		var res = [1, 2, 3].map(fn);
-		order.push('sync');
-
-		return res[2].then(function(res) {
-			expect(res).to.equal(30);
-			expect(order).to.deep.equal(['sync', 'start1', 'end1', 'start2', 'end2', 'start3', 'end3']);
-		});
-	});
-
-	it('passes this context', function() {
-		var fn = series(function*(i) {
-			yield Promise.resolve();
-			return i + '-' + this.j;
-		});
-
-		var x = {fn: fn, j: 9};
-
-		return x.fn(3).then(function(res) {
-			expect(res).to.equal('3-9');
-		});
-	});
-});
+runTests(order,
+	function*(i) {
+		order.push('start' + i);
+		yield Promise.resolve();
+		order.push('end' + i);
+		return i * 10;
+	},
+	function*(i) {
+		yield Promise.resolve();
+		return i + '-' + this.j;
+	},
+	function*() {
+		yield Promise.resolve();
+	}
+);